requirements.txt
altgraph==0.17.4
packaging==24.2
pillow==11.1.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
qrcode==8.0
setuptools==75.8.0
